Add support in mapr for setting target-specific pipeline config
The target-specific p4info and device config blob (e.g., bmv2.json) can
be passed as command line flags. When a P4RT controller sends a
SetForwardingPipelineConfig, mapr verifies that the request's P4Info is
equal to a known logical one (also passed via command line flags), if
true, it modifies the request with the target-specific config.
Otherwise, it returns an error.

For convenience, we check in target-specific configs such as for
fabric.p4 (p4info.bin and bmv2.json).