First stab at generic translation logic based on target-specific change processors (#13)
We add initial infrastructure to support translating P4RT for fabric.p4 and other targets.

This PR adds:

TassenStore, a store with APIs to manipulate the state of high-level logical objects (e.g, IfTypeEntry, AttachmentEntry, etc.). This store is updated as a consequence of P4RT Write RPCs for the logical pipeline but offers a more abstract representation than P4Runtime entries, convenient for the translator operations.
ChangeProcessor interface, introduced to abstract commonality in the translation logic between different physical pipelines.
Skeleton implementation of ChangeProcessor for fabric.p4.
An implementation of a generic Translator based on ChangeProcessor.
The translation logic is based on the following rationale.

P4 tables and other objects can be categorized based on the type of information they hold, which can be device-level, or attachment-level. For updates to device-level objects, we parse them and call the corresponding method in the change processor. For attachment-level updates, the state of an attachment depends on many tables, as such, a translator might need to "accumulate" updates before it can modify the physical pipeline. When receiving an update for an attachment-level table, we first evaluate a snapshot of the attachment that includes data from TassenStore and from the given update. The snapshot is passed to the change processor with an indication of whether the attachment state is complete (all fields are populated) or not.

Considering an example implementation for a change processor, if an attachment is complete, the processor might generate all necessary updates to insert the corresponding entries in the target to enable termination/forwarding. If the attachment is not complete, the processor might decide to remove all entries from the target related to that attachment.

The implementation is still work in progress. As of now, it supports only upstream tables.