Support only CONTINUE_ON_ERROR atomicity for WriteRequest
That allows us to simplify the translator and store implementation, by
evaluating logical updates one-by-one, rather than in batches. As a
consequence, given a logical WriteRequest with N updates, the new
implementation might generate up to N physical WriteRequests for the
target. That's bad for performance, but for now it makes it easier to
reason about the translation logic and the stores.