Simplify downstream pipeline
Remove ECMP and merge all attachments attributes in one table, including
port and MAC address (previously hold in the routing table).

This design should facilitate a simpler translation logic in mapr.

ECMP is not needed as the assumption is that the attachment is reachable
via an L2 tunnel. There might be a need for LAGs, and hence hashing, but
that is different from ECMP. LAGs can be handled by assuming logical port
numbers mapped to a lAG group in some other stage of the pipeline, which
for now is out of scope for this work.